seeders

+++ app/Models/Film.php
@@ -14,6 +14,7 @@ class Film extends Model
     use HasFactory;
 
     protected $fillable = ['title', 'year', 'genre', 'image'];
+    
 
     public function subtitles()
     {
@@ -21,36 +22,40 @@ class Film extends Model
     }
 
     public static function boot()
-{
-    parent::boot();
-    static::deleting(function($obj) {
-        Storage::delete(Str::replaceFirst('storage/','public/', $obj->image));
-    });
-}
+    {
+        parent::boot();
+    
+        static::deleting(function($obj) {
+            if ($obj->image !== null) {
+                Storage::delete(Str::replaceFirst('storage/','public/', $obj->image));
+            }
+        });
+    }
 
 
 
 public function setImageAttribute($value)
 {
+
     $attribute_name = "image";
-    // destination path relative to the disk above
     $destination_path = "films";
 
-    // if the image was erased
-    if ($value==null) {
-        // delete the image from disk
-        Storage::delete(Str::replaceFirst('storage/','public/',$this->{$attribute_name}));
+    if ($value == null) {
+        Storage::delete(Str::replaceFirst('storage/', 'public/', $this->{$attribute_name}));
 
-        // set null in the database column
         $this->attributes[$attribute_name] = null;
-    }
-
-    $disk = "public";
-    // filename is generated -  md5($file->getClientOriginalName().random_int(1, 9999).time()).'.'.$file->getClientOriginalExtension()
-    $this->uploadFileToDisk($value, $attribute_name, $disk, $destination_path, $fileName = null);
-    $this->attributes[$attribute_name] = 'storage/' . $this->attributes[$attribute_name];
-
+    } elseif ($value instanceof \Illuminate\Http\UploadedFile) {
+        $disk = "public";
+        $fileName = md5($value->getClientOriginalName() . random_int(1, 9999) . time()) . '.' . $value->getClientOriginalExtension();
+        Storage::putFileAs("public/" . $destination_path, $value, $fileName, $disk);
+        $this->attributes[$attribute_name] =  'storage/films/' . $fileName;
+    } else {
+        $this->attributes[$attribute_name] = 'storage/films/' . $value;
 }
+}
+
+
+
 
 public function scopeSearch($query, $search)
 {
